requirements.txt
# Django Framework
Django>=5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-filter>=23.0
django-extensions>=3.2.0
django-guardian>=2.4.0

# Database
psycopg2-binary>=2.9.9

# Authentication
djangorestframework-simplejwt>=5.3.0
PyJWT>=2.8.0

# Celery & Redis
celery>=5.3.0
redis>=5.0.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.0

# File Handling
Pillow>=10.0.0
django-storages>=1.14.0

# Security
python-dotenv>=1.0.0
whitenoise>=6.6.0
argon2-cffi>=23.1.0

# Payments
stripe>=7.8.0

# Email
django-anymail>=10.2
sendgrid>=6.11.0

# Documentation
drf-spectacular>=0.26.0

# Testing
pytest>=7.4.0
pytest-django>=4.5.0
pytest-cov>=4.1.0
factory-boy>=3.3.0
faker>=20.0.0

# Development
ipdb>=0.13.13
django-debug-toolbar>=4.2.0

# Production Server
gunicorn>=21.2.0
uvicorn>=0.25.0

# Utilities
python-dateutil>=2.8.2
requests>=2.31.0
